Conflicts:
	EagleEventPlanning-HTML/guestList.html
	
	Got printing working in Firfox, doesn't work in Edge.  not sure about
Chrome. updated clientlist to start with first client selected

+++ EagleEventPlanning/src/ProblemDomain/EventSystem.java
@@ -109,7 +109,7 @@ public class EventSystem {
 	}
 
 	public static Guest findGuestById(String id) {
-		return GuestDAO.findGuestById(Long.parseLong(id));
+		return GuestDAO.findGuestById(Integer.parseInt(id));
 	}
 
 	public static Boolean removeGuest(Guest guest) {
